SliderItems.tsx has been added to mock-data from home.tsx
